+++ Assets/Scripts/Entities/Enemy/AIController.cs
@@ -29,13 +29,13 @@ public class AIController : MonoBehaviour
     private bool knowAboutPlayer;
     private bool fallBack;
 
-    private int buffDebuffDmg = 0;
+    private float buffDebuffDmg = 0;
 
     private float elapsedTime = 0;
 
     private float savedSpeedToRestore;
 
-    public void ChangeBuffDebuffDmgBy(int value)
+    public void ChangeBuffDebuffDmgBy(float value)
     {
         buffDebuffDmg += value;
     }
@@ -242,7 +242,7 @@ public class AIController : MonoBehaviour
         fullDistance = temp;
     }
 
-    public int getDamage()
+    public float getDamage()
     {
         return enemyData.lightAttackDamage + buffDebuffDmg;
     }
