+++ Assets/Scripts/Items/Weapons/WeaponData.cs
@@ -23,6 +23,6 @@ public class WeaponData : ScriptableObject
     public string bashAnimation;
     public string hurtAnimation;
     [Header("Weapon stats")]
-    public int lightAttackDamage;
-    public int heavyAttackDamage;
+    public float lightAttackDamage;
+    public float heavyAttackDamage;
 }
